+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   - `str(Operation)` is no longer required to be unique (except with the ZI backend).
   - `schedule.operations.keys()` can no longer be used to get a repr of the Operations.
 - Qblox backend - Make q1asm generation functions `acquire_append` and `acquire_average` private. (!739)
+- Qblox backend - `to_grid_time` helper function would raise `ValueError` if time supplied is not a multiple of grid time, now additionally checking that time is within tolerance of 1 picosecond of grid time instead of silently rounding to nanoseconds (!751)
+- Qblox backend - Rename `is_within_grid_time ` helper function to `is_within_half_grid_time` !(753)
 
 ### Merged branches and closed issues
 
@@ -26,7 +28,6 @@
 - Waveforms - Fix such that `interpolated_complex_waveform` does not extrapolate except for rounding errors. (!710)
 - Docs - Fix broken list bullets on Read-the-Docs by enforcing `sphinx-rtd-theme>=1.2.2` (!743)
 - Qblox backend - Added warning if waveform playback or acquisition is interrupted by another waveform or acquisition, respectively (!744, #436)
-- Qblox backend - `to_grid_time` helper function extended to alert user in case of time value not an integer multiple of nanosecond (tolerance 1 picosecond) (!751)
 
 ## 0.15.0 (2023-07-13)